fix(admin+mobile): resolve order creation, detail view, and dashboard stats bugs
- fix(api): make id_equipo optional in CreateOrdenDto for multi-equipment support
- fix(api): transform ordenes_equipos.equipos -> equipo in query handlers
- fix(mobile): dashboard now counts ASIGNADA, PROGRAMADA, EN_PROCESO, EN_ESPERA_REPUESTO as active orders
- docs: document all bug fixes in CHAT file

+++ apps/mobile/lib/core/sync/sync_service.dart
@@ -548,7 +548,7 @@ class SyncService {
             // ✅ NUEVO: Guardar equipos de la orden (multi-equipos)
             final equiposData = orden['ordenesEquipos'] as List?;
             debugPrint('🔧 [SYNC] Orden ${orden['numeroOrden']} (ID $idOrden) - ordenesEquipos recibido del backend: ${equiposData?.length ?? 0}');
-            await _guardarOrdenesEquipos(idOrden as int, equiposData);
+            await _guardarOrdenesEquipos(idOrden, equiposData);
           } else {
             final idLocalNueva = await _db.insertOrdenFromSync(ordenCompanion);
 
@@ -563,7 +563,7 @@ class SyncService {
             // ✅ NUEVO: Guardar equipos de la orden (multi-equipos)
             final equiposDataNueva = orden['ordenesEquipos'] as List?;
             debugPrint('🔧 [SYNC] Orden NUEVA ${orden['numeroOrden']} (ID $idOrden) - ordenesEquipos recibido del backend: ${equiposDataNueva?.length ?? 0}');
-            await _guardarOrdenesEquipos(idOrden as int, equiposDataNueva);
+            await _guardarOrdenesEquipos(idOrden, equiposDataNueva);
           }
           ordenesCount++;
         } catch (e) {
